Move sources to internal package
