+++ internal/newsapi/client.go
