+++ internal/newsapi/errors.go
