+++ internal/newsapi/service.go
